+++ CHANGELOG.rst
@@ -12,6 +12,16 @@ The semantic versioning only considers the public API as described in
 paths are considered internals and can change in minor and patch releases.
 
 
+v4.50.1 (unreleased)
+--------------------
+
+Fixed
+^^^^^
+- ``TypeError`` when adding arguments for a class that has a parameter that has
+  a default and is annotated with a non-``@runtime_checkable`` ``Protocol`` type
+  (`#935 <https://github.com/mauvilsa/jsonargparse/pull/935>`__).
+
+
 v4.50.0 (2026-07-22)
 --------------------